Add implementation for structure builder
This changes the relevant blocks and spawns in the relevant block
display entities to create the waystone structure, as it was originally
implemented in the presentation layer.
