+++ src/main/kotlin/dev/mizarc/waystonewarps/application/services/StructureBuilderService.kt
@@ -1,5 +1,7 @@
 package dev.mizarc.waystonewarps.application.services
 
+import dev.mizarc.waystonewarps.domain.warps.Warp
+
 /**
  * Service that handles the building of custom waystone structures.
  */
@@ -7,10 +9,10 @@ interface StructureBuilderService {
     /**
      * Builds the waystone structure in the world.
      */
-    fun spawnStructure()
+    fun spawnStructure(warp: Warp)
 
     /**
      * Destroys the waystone structure in the world.
      */
-    fun despawnStructure()
+    fun despawnStructure(warp: Warp)
 }
